pyproject.toml
[project]
name = "surplus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.4"
dependencies = [
    "numpy>=2.2.0",
    "pyqt6>=6.7.1",
    "pyyaml>=6.0.2",
]
